junk

+++ Minirel/bufHash.C
@@ -13,7 +13,7 @@
 int BufHashTbl::hash(const File* file, const int pageNo)
 {
   int tmp, value;
-  tmp = (long)file;  // cast of pointer to the file object to an integer
+  tmp = (int)(long)file;  // cast of pointer to the file object to an integer
   value = (tmp + pageNo) % HTSIZE;
   return value;
 }
